Added the hint button and a label to display messages to the user.
The hint button and label can be found at the bottom of the gui. A user can press the hint button to enable a private boolean variable called showPath. The idea is that if showPath is enabled then, whenever we draw the maze, the path can be shown the user. The code is such that the user can only see the path for a certain number of movements before a cooldown is set. Cooldown is currently set to 5 player movements.